+++ test/components/alg-iron-checked-element-behavior/components.js
@@ -1,5 +1,5 @@
 import '../demo-pages.js';
 import './style.js';
-import './simple-checkbox.js';
 
 document.body.style.opacity = null;
+import './simple-checkbox.js';