+++ test/components/alg-iron-dropdown/components.js
@@ -1,8 +1,8 @@
 import '../demo-pages.js';
 import './style.js';
+document.body.style.opacity = null;
+
 import { AlgIronDropdown } from '../../../lib/components/alg-iron-dropdown.js';
 import { AlgIronImage } from '../../../lib/components/alg-iron-image.js';
 import './x-select.js';
 import './x-ul.js';
-
-document.body.style.opacity = null;