+++ test/components/alg-iron-icon/components.js
@@ -1,6 +1,6 @@
 import '../../../lib/import-all-icons.js';
 import '../demo-pages.js';
+document.body.style.opacity = null;
+
 import { AlgIronIcon } from '../../../lib/components/alg-iron-icon.js';
 import { AppDemo } from './app-demo.js';
-
-document.body.style.opacity = null;