+++ test/components/alg-iron-resizable-behavior/components.js
@@ -1,5 +1,5 @@
 import '../demo-pages.js';
 import './style.js';
-import './x-app.js';
-
 document.body.style.opacity = null;
+
+import './x-app.js';