+++ test/components/alg-iron-selector/components.js
@@ -1,8 +1,7 @@
 import '../demo-pages.js';
 import './style.js';
+document.body.style.opacity = null;
 
 import { AlgIronSelector } from '../../../lib/components/alg-iron-selector.js';
 import { AlgPaperRadioButton } from '../../../lib/components/alg-paper-radio-button.js';
 import { AlgPaperToggleButton } from '../../../lib/components/alg-paper-toggle-button.js';
-
-document.body.style.opacity = null;