+++ test/components/alg-iron-validatable-behavior/components.js
@@ -1,7 +1,6 @@
 import '../demo-pages.js';
 import './style.js';
+document.body.style.opacity = null;
 
 import './cats-only.js'; // load validator
 import './validatable-input.js';
-
-document.body.style.opacity = null;