+++ test/components/alg-paper-button/components.js
@@ -1,7 +1,8 @@
-import '../../../lib/icons/alg-iron-icons.js';
 import '../demo-pages.js';
 import './style.js';
+document.body.style.opacity = null;
+
+import '../../../lib/icons/alg-iron-icons.js';
 import { AlgIronIcon } from '../../../lib/components/alg-iron-icon.js';
 import { AlgPaperButton } from '../../../lib/components/alg-paper-button.js';
-
-document.body.style.opacity = null;
+import '../show-msg.js';