+++ test/components/alg-paper-fab/components.js
@@ -1,6 +1,7 @@
-import '../../../lib/icons/alg-iron-icons.js';
 import '../demo-pages.js';
 import './style.js';
-import { AlgPaperFab } from '../../../lib/components/alg-paper-fab.js';
-
 document.body.style.opacity = null;
+
+import '../../../lib/icons/alg-iron-icons.js';
+import { AlgPaperFab } from '../../../lib/components/alg-paper-fab.js';
+import '../show-msg.js';