+++ test/components/alg-paper-icon-button/components.js
@@ -1,6 +1,7 @@
-import '../../../lib/icons/alg-iron-icons.js';
 import '../demo-pages.js';
 import './style.js';
-import { AlgPaperIconButton } from '../../../lib/components/alg-paper-icon-button.js';
-
 document.body.style.opacity = null;
+
+import '../../../lib/icons/alg-iron-icons.js';
+import { AlgPaperIconButton } from '../../../lib/components/alg-paper-icon-button.js';
+import '../show-msg.js';