+++ test/components/alg-paper-radio-group/components.js
@@ -1,7 +1,7 @@
 import '../demo-pages.js';
 import './style.js';
+document.body.style.opacity = null;
 
 import { AlgPaperRadioButton } from '../../../lib/components/alg-paper-radio-button.js';
 import { AlgPaperRadioGroup } from '../../../lib/components/alg-paper-radio-group.js';
-
-document.body.style.opacity = null;
+import '../show-msg.js';